backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
neo4j==5.14.1
py2neo==2021.2.4

# Machine Learning
torch==2.1.1
torch-geometric==2.4.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# NLP
spacy==3.7.2
transformers==4.36.2

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For TestClient async support
